requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openai==1.76.0
chromadb==0.6.3
flask-cors==5.0.0
pydantic==2.8.2
typing-extensions==4.12.2
python-docx
PyMuPDF

